+++ test/fixtures/go-module/demo_test.go
@@ -1,0 +1,5 @@
+package demo
+
+import "testing"
+
+func TestDemo(t *testing.T) {}